Restructure to follow three teir architecture
