+++ lib/decorators/capitalize_decorator.rb
