+++ lib/decorators/decorator.rb
