+++ lib/decorators/nameable.rb
