+++ lib/decorators/trimmer_decorator.rb
