+++ lib/domain/book.rb
