+++ lib/domain/class_room.rb
