+++ lib/domain/person.rb
