+++ lib/domain/rental.rb
