+++ lib/domain/student.rb
