+++ lib/domain/teacher.rb
